+++ src/core/operations/BcryptParse.mjs
@@ -20,7 +20,7 @@ class BcryptParse extends Operation {
         super();
 
         this.name = "Bcrypt parse";
-        this.module = "Hashing";
+        this.module = "Crypto";
         this.description = "Parses a bcrypt hash to determine the number of rounds used, the salt, and the password hash.";
         this.infoURL = "https://wikipedia.org/wiki/Bcrypt";
         this.inputType = "string";