+++ src/core/operations/CompareCTPHHashes.mjs
@@ -22,7 +22,7 @@ class CompareCTPHHashes extends Operation {
         super();
 
         this.name = "Compare CTPH hashes";
-        this.module = "Hashing";
+        this.module = "Crypto";
         this.description = "Compares two Context Triggered Piecewise Hashing (CTPH) fuzzy hashes to determine the similarity between them on a scale of 0 to 100.";
         this.infoURL = "https://forensicswiki.org/wiki/Context_Triggered_Piecewise_Hashing";
         this.inputType = "string";